backend/pyproject.toml
[tool.poetry]
name = "graintrade-api"
version = "0.1.0"
description = "Graintrade API"
authors = ["Kostiantyn Ivashchenko <kivaschenko@protonmail.com>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
uvicorn = "^0.32.0"
gunicorn = "^23.0.0"
PyJWT = "^2.9.0"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
bcrypt = "^4.2.0"
psycopg2-binary = "^2.9.10"
aio-pika = "^9.5.5"
httpx = "^0.28.1"
redis = {extras = ["async"], version = "^5.2.1"}
orjson = "^3.10.16"
requests = "^2.32.3"
fastapi-mail = "^1.5.0"
dotenv = "^0.9.9"
aiohttp = "^3.9.0"
prometheus-client = "^0.21.0"
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
